refactor(deposits): extract usePendingDeposits from the unclaimed-deposit banner
The generation-guarded fetch (focus + wallet-refresh triggers) moves into a
shared hook so the balance header can read pending deposits too. The hook
skips setState when the fetched list is unchanged, keeping the array
identity stable for effects keyed on it. Banner behavior is unchanged and
its spec passes as-is.
