+++ app/self-custodial/hooks/index.ts
@@ -1,5 +1,6 @@
 export { useNonCustodialConversionLimits } from "./use-non-custodial-conversion-limits"
 export { usePaymentRequest } from "./use-payment-request"
+export { usePendingDeposits } from "./use-pending-deposits"
 export { useSendDustWarning } from "./use-send-dust-warning"
 export { useTranslateSdkError } from "./use-translate-sdk-error"
 export type { SelfCustodialPaymentRequestState, InvoiceData } from "./types"
